lambda/pokemon_events_generator/requirements.txt
kafka-python==2.0.2
faker